refactor: separate new-task
